pipeline/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
